bug: handle subpixel values in viewport dims
Uses the bounding client rect of the `documentElement` for getting the
width / height of the viewport instead of the jQuery `innerWidth` /
`innerHeight` values. The former returns subpixel accuracy while the
latter rounds to whole numbers.

This is particularly important when the user has zoomed the window,
in which case the browser may scale the size of the viewport, resulting
in the dimensions having fractional parts, e.g., the width might be
`1451.1112060546875`.

Closes #96